Update open IP address rule
As Azure has a update about 'enable access to all the Azure serivces' which is equal to 0.0.0.0 starting IP and 0.0.0.0 as ending IP. So we cannot add this in the firewall rule as temporary testing purposes. 

The right way to open the server to all IP addresses is to set up a firewall rule using 0.0.0.0 as the starting IP address range and using 255.255.255.255 as the ending IP address range. Although, it is not recommended.